move mixin into better place, make lakes generate earlier then trees and foliage, fix worldtype stuff
